Moved config to startup and cleaned filesystem
