+++ RepairsApi.Tests/V2/UseCase/ListWorkOrdersUseCaseTests.cs
@@ -6,6 +6,7 @@ using RepairsApi.Tests.V2.Gateways;
 using RepairsApi.V2.Boundary.Response;
 using RepairsApi.V2.Controllers.Parameters;
 using RepairsApi.V2.Factories;
+using RepairsApi.V2.Filtering;
 using RepairsApi.V2.Infrastructure;
 using RepairsApi.V2.Infrastructure.Extensions;
 using RepairsApi.V2.UseCase;
@@ -28,7 +29,27 @@ namespace RepairsApi.Tests.V2.UseCase
         {
             configureGenerator();
             _repairsMock = new MockRepairsGateway();
-            _classUnderTest = new ListWorkOrdersUseCase(_repairsMock.Object);
+            _classUnderTest = new ListWorkOrdersUseCase(_repairsMock.Object, CreateFilterBuilder());
+        }
+
+        private static FilterBuilder<WorkOrderSearchParameters, WorkOrder> CreateFilterBuilder()
+        {
+            return new FilterBuilder<WorkOrderSearchParameters, WorkOrder>()
+                .AddFilter(
+                    searchParams => searchParams.ContractorReference,
+                    contractorReference => !string.IsNullOrWhiteSpace(contractorReference),
+                    contractorReference => wo => wo.AssignedToPrimary.ContractorReference == contractorReference
+                )
+                .AddFilter(
+                    searchParams => searchParams.PropertyReference,
+                    p => !string.IsNullOrWhiteSpace(p),
+                    p => wo => wo.Site.PropertyClass.Any(pc => pc.PropertyReference == p)
+                )
+                .AddFilter(
+                    searchParams => searchParams.StatusCode,
+                    code => code > 0 && Enum.IsDefined(typeof(WorkStatusCode), code),
+                    code => wo => wo.StatusCode == (WorkStatusCode) Enum.Parse(typeof(WorkStatusCode), code.ToString())
+                );
         }
 
         private void configureGenerator()
@@ -158,7 +179,7 @@ namespace RepairsApi.Tests.V2.UseCase
 
             var generatedWorkOrders = _generator.GenerateList(workOrderCount);
 
-            _repairsMock.Setup(r => r.GetWorkOrders())
+            _repairsMock.Setup(r => r.GetWorkOrders(It.IsAny<IFilter<WorkOrder>>()))
                 .ReturnsAsync(generatedWorkOrders);
             return generatedWorkOrders;
         }
@@ -174,7 +195,7 @@ namespace RepairsApi.Tests.V2.UseCase
 
             var generatedWorkOrders = generator.GenerateList(workOrderCount);
 
-            _repairsMock.Setup(r => r.GetWorkOrders())
+            _repairsMock.Setup(r => r.GetWorkOrders(It.IsAny<IFilter<WorkOrder>>()))
                 .ReturnsAsync(generatedWorkOrders);
             return generatedWorkOrders;
         }
