+++ RepairsApi/V2/Filtering/IFilter.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+using System.Linq;
+
+namespace RepairsApi.V2.Filtering
+{
+    public interface IFilter<TQuery>
+    {
+        IQueryable<TQuery> Apply(IQueryable<TQuery> query);
+        IEnumerable<TQuery> Apply(IEnumerable<TQuery> query);
+    }
+}
