+++ RepairsApi/V2/Filtering/IFilterBuilder.cs
@@ -1,0 +1,7 @@
+namespace RepairsApi.V2.Filtering
+{
+    public interface IFilterBuilder<TSearch, TQuery>
+    {
+        IFilter<TQuery> BuildFilter(TSearch searchParameter);
+    }
+}
